New java

+++ Java/CodeWithHarry/CWH_02_Quiz.java
@@ -5,7 +5,7 @@ public class CWH_02_Quiz {
     public static void main(String[] args) {
 
         int sum = 0;
-        int percent = 0;
+        float percent = 0;
         int mark = 0;
         Scanner sc = new Scanner(System.in);
         for (int i = 1; i <= 5; i++) {
@@ -15,7 +15,7 @@ public class CWH_02_Quiz {
             sum += mark;
         }
         // System.out.println((floa);
-        percent = (sum / 500) * 100;
+        percent = ((sum / 500.0f) * 100);
         System.out.println("Percent " + (float) percent);
         sc.close();
